+++ src/main/java/de/dealog/msg/messaging/tracking/MessagesTrackingPayload.java
@@ -1,8 +1,8 @@
 package de.dealog.msg.messaging.tracking;
 
 import io.quarkus.runtime.annotations.RegisterForReflection;
+import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 import lombok.ToString;
 
 import java.util.Date;
@@ -11,8 +11,8 @@ import java.util.List;
 /**
  * The payload contains the data of the message
  */
+@Builder
 @Getter
-@Setter
 @ToString
 @RegisterForReflection
 public class MessagesTrackingPayload {
@@ -20,12 +20,12 @@ public class MessagesTrackingPayload {
     /**
      * A list of tracked message ids
      */
-    private List<String> ids;
+    private final List<String> ids;
 
     /**
      * The date and time when the message request was tracked in UTC
      */
-    private Date trackedAt;
+    private final Date trackedAt;
 }