+++ src/main/java/de/dealog/msg/rest/RegionResourceConstants.java
@@ -1,0 +1,26 @@
+package de.dealog.msg.rest;
+
+/**
+ * Constants used by the {@link RegionResource}
+ */
+public final class RegionResourceConstants{
+    /**
+     * Regions path
+     */
+    static final String RESOURCE_PATH = "/api/regions";
+
+    /**
+     * URI path for hierarchy
+     */
+    static final String PATH_HIERACHY = "hierarchy";
+
+    /**
+     * URI template parameter for ars
+     */
+    static final String PATH_PARAM_ARS = "ars";
+
+    /**
+     * Query parameter for name
+     */
+    static final String QUERY_NAME = "name";
+}