+++ src/test/java/de/dealog/msg/rest/RegionResourceTest.java
@@ -1,13 +1,14 @@
 package de.dealog.msg.rest;
 
 import de.dealog.msg.TestUtils;
+import de.dealog.msg.converter.RegionConverter;
 import de.dealog.msg.persistence.model.Region;
 import de.dealog.msg.rest.model.GeoRequest;
 import de.dealog.msg.rest.model.PageRequest;
-import de.dealog.msg.rest.model.PagedList;
 import de.dealog.msg.rest.model.RegionTypeRest;
 import de.dealog.msg.rest.validations.ValidGeoRequest;
 import de.dealog.msg.service.RegionService;
+import de.dealog.msg.service.model.PagedList;
 import de.dealog.msg.service.model.QueryParams;
 import io.quarkus.test.junit.QuarkusMock;
 import io.quarkus.test.junit.QuarkusTest;
@@ -18,8 +19,6 @@ import org.mockito.Mockito;
 import javax.inject.Inject;
 import java.util.Arrays;
 
-import static de.dealog.msg.TestUtils.SIZE_FAILS;
-import static de.dealog.msg.rest.RegionResource.*;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.*;
 import static org.mockito.ArgumentMatchers.any;
@@ -52,8 +51,8 @@ class RegionResourceTest {
     @Test
     void findHierachy() {
        given()
-               .param(PATH_ARS, "091790134135")
-               .when().get(RESOURCE_PATH + "/" + PATH_HIERACHY)
+               .param(RegionResourceConstants.PATH_PARAM_ARS, "091790134135")
+               .when().get(RegionResourceConstants.RESOURCE_PATH + "/" + RegionResourceConstants.PATH_HIERACHY)
                .then()
                 .statusCode(200)
                 .body(startsWith("[{\"ars\":\"000000000000\""))
@@ -69,7 +68,7 @@ class RegionResourceTest {
                 .param("ars", "0")
                 .param(PageRequest.PAGE, 0)
                 .param(PageRequest.SIZE, 10)
-                .when().get(RESOURCE_PATH + "/" + PATH_HIERACHY)
+                .when().get(RegionResourceConstants.RESOURCE_PATH + "/" + RegionResourceConstants.PATH_HIERACHY)
                 .then()
                 .statusCode(400)
                 .body(containsString(TestUtils.SIZE_FAILS));
@@ -81,7 +80,7 @@ class RegionResourceTest {
                 .param("ars", "0917901341345")
                 .param(PageRequest.PAGE, 0)
                 .param(PageRequest.SIZE, 10)
-                .when().get(MessageResource.RESOURCE_PATH)
+                .when().get(MessageConstants.RESOURCE_PATH)
                 .then()
                 .statusCode(400)
                 .body(containsString(TestUtils.SIZE_FAILS));
@@ -93,7 +92,7 @@ class RegionResourceTest {
                 .param("ars", "0a1b2c3d4f")
                 .param(PageRequest.PAGE, 0)
                 .param(PageRequest.SIZE, 10)
-                .when().get(RESOURCE_PATH + "/" + PATH_HIERACHY)
+                .when().get(RegionResourceConstants.RESOURCE_PATH + "/" + RegionResourceConstants.PATH_HIERACHY)
                 .then()
                 .statusCode(400)
                 .body(containsString(TestUtils.PATTERN_FAILS));
@@ -103,14 +102,14 @@ class RegionResourceTest {
     void findHierachy_LatAndLongFails() {
         given()
                 .param(GeoRequest.LATITUDE, 48.21667)
-                .when().get(RESOURCE_PATH + "/" + PATH_HIERACHY)
+                .when().get(RegionResourceConstants.RESOURCE_PATH + "/" + RegionResourceConstants.PATH_HIERACHY)
                 .then()
                 .statusCode(400)
                 .body(containsString(ValidGeoRequest.MESSAGE));
 
         given()
                 .param(GeoRequest.LONGITUDE, 11.26667)
-                .when().get(RESOURCE_PATH + "/" + PATH_HIERACHY)
+                .when().get(RegionResourceConstants.RESOURCE_PATH + "/" + RegionResourceConstants.PATH_HIERACHY)
                 .then()
                 .statusCode(400)
                 .body(containsString(ValidGeoRequest.MESSAGE));